+++ app/assets/javascripts/pageflow/editor/views/confirmable_file_item_view.js
@@ -1,0 +1,42 @@
+pageflow.ConfirmableFileItemView = Backbone.Marionette.ItemView.extend({
+  tagName: 'li',
+  template: 'templates/confirmable_file_item',
+
+  ui: {
+    fileName: '.file_name',
+    duration: '.duration',
+
+    label: 'label',
+    checkBox: 'input',
+    removeButton: '.remove',
+  },
+
+  events: {
+    'click .remove': 'destroy',
+
+    'change input': 'updateSelection'
+  },
+
+  onRender: function() {
+    this.ui.label.attr('for', this.cid);
+    this.ui.checkBox.attr('id', this.cid);
+
+    this.ui.fileName.text(this.model.get('file_name') || '(Unbekannt)');
+    this.ui.duration.text(this.model.get('duration') || '-');
+  },
+
+  destroy: function() {
+    if (confirm("Datei wirklich wirklich löschen?")) {
+      this.model.destroyUsage();
+    }
+  },
+
+  updateSelection: function() {
+    if (this.ui.checkBox.is(':checked')) {
+      this.options.selectedFiles.add(this.model);
+    }
+    else {
+      this.options.selectedFiles.remove(this.model);
+    }
+  },
+});