+++ app/assets/javascripts/pageflow/editor/views/notifications_view.js
@@ -5,7 +5,8 @@ pageflow.NotificationsView = Backbone.Marionette.ItemView.extend({
 
   ui: {
     failedCount: '.failed .count',
-    uploadingCount: '.uploading .count'
+    uploadingCount: '.uploading .count',
+    confirmableFilesCount: '.confirmable_files .count',
   },
 
   events: {
@@ -16,12 +17,15 @@ pageflow.NotificationsView = Backbone.Marionette.ItemView.extend({
 
   onRender: function() {
     this.listenTo(pageflow.entry, 'change:uploading_files_count', this.notifyUploadCount);
+    this.listenTo(pageflow.entry, 'change:confirmable_files_count', this.notifyConfirmableFilesCount);
+
     this.listenTo(pageflow.failedRecords, 'add', this.update);
     this.listenTo(pageflow.failedRecords, 'remove', this.update);
     this.listenTo(pageflow.savingRecords, 'add', this.update);
     this.listenTo(pageflow.savingRecords, 'remove', this.update);
 
     this.update();
+    this.notifyConfirmableFilesCount();
   },
 
   update: function() {
@@ -33,5 +37,12 @@ pageflow.NotificationsView = Backbone.Marionette.ItemView.extend({
   notifyUploadCount: function(model, uploadCount) {
     this.$el.toggleClass('uploading', uploadCount > 0);
     this.ui.uploadingCount.text(uploadCount);
+  },
+
+  notifyConfirmableFilesCount: function() {
+    var confirmableFilesCount = pageflow.entry.get('confirmable_files_count');
+
+    this.$el.toggleClass('has_confirmable_files', confirmableFilesCount > 0);
+    this.ui.confirmableFilesCount.text(confirmableFilesCount);
   }
 });