+++ app/jobs/pageflow/poll_meta_data_from_zencoder_job.rb
@@ -1,0 +1,9 @@
+module Pageflow
+  class PollMetaDataFromZencoderJob < PollZencoderJob
+    @queue = :resizing
+
+    def self.perform_with_result(file, options)
+      super(file, :skip_thumbnail => true)
+    end
+  end
+end