+++ app/jobs/pageflow/request_meta_data_from_zencoder_job.rb
@@ -1,11 +1,11 @@
 module Pageflow
-  class SubmitFileToZencoderJob
+  class RequestMetaDataFromZencoderJob
     @queue = :default
 
     extend StateMachineJob
 
     def self.perform_with_result(file, options, api = ZencoderApi.instance)
-      file.job_id = api.create_job(file.output_definition)
+      file.job_id = api.create_job(ZencoderMetaDataOutputDefinition.new(file))
       file.save!
 
       :ok