+++ app/state_machines/pageflow/encoded_file_state_machine.rb
@@ -8,11 +8,15 @@ module Pageflow
 
         state 'not_uploaded_to_s3'
         state 'uploading_to_s3'
+        state 'waiting_for_meta_data'
+        state 'fetching_meta_data'
+        state 'waiting_for_confirmation'
         state 'waiting_for_encoding'
         state 'encoding'
         state 'encoded'
 
         state 'upload_to_s3_failed'
+        state 'fetching_meta_data_failed'
         state 'encoding_failed'
 
         event :publish do
@@ -21,6 +25,10 @@ module Pageflow
 
         event :retry do
           transition 'upload_to_s3_failed' => 'uploading_to_s3'
+
+          transition 'encoding_failed' => 'waiting_for_confirmation', :if => lambda { Pageflow.config.confirm_encoding_jobs }
+          transition 'encoded' => 'waiting_for_confirmation', :if => lambda { Pageflow.config.confirm_encoding_jobs }
+
           transition 'encoding_failed' => 'waiting_for_encoding'
           transition 'encoded' => 'waiting_for_encoding'
         end
@@ -28,8 +36,32 @@ module Pageflow
         job UploadFileToS3Job do
           on_enter 'uploading_to_s3'
           result :pending, :retry_after => 30.seconds
-          result :ok => 'waiting_for_encoding'
+
+          result :ok, :state => 'waiting_for_meta_data', :if => lambda { Pageflow.config.confirm_encoding_jobs }
+          result :ok, :state => 'waiting_for_encoding'
+
           result :error => 'uploading_to_s3_failed'
+        end
+
+        job RequestMetaDataFromZencoderJob do
+          on_enter 'waiting_for_meta_data'
+          result :ok => 'fetching_meta_data'
+          result :error => 'fetching_meta_data_failed'
+        end
+
+        job PollMetaDataFromZencoderJob do
+          on_enter 'fetching_meta_data'
+          result :pending, :retry_after => 2.seconds
+          result :ok => 'waiting_for_confirmation'
+          result :error => 'fetching_meta_data_failed'
+        end
+
+        event :confirm_encoding do
+          transition 'waiting_for_confirmation' => 'waiting_for_encoding'
+        end
+
+        after_transition 'waiting_for_confirmation' => 'waiting_for_encoding' do |file|
+          Pageflow.config.hooks.invoke(:file_encoding_confirmed, :file => file)
         end
 
         job SubmitFileToZencoderJob do
@@ -43,6 +75,14 @@ module Pageflow
           result :pending, :retry_after => 2.seconds
           result :ok => 'encoded'
           result :error => 'encoding_failed'
+        end
+
+        after_transition any => 'encoded' do |file|
+          Pageflow.config.hooks.invoke(:file_encoded, :file => file)
+        end
+
+        after_transition any => 'encoding_failed' do |file|
+          Pageflow.config.hooks.invoke(:file_encoding_failed, :file => file)
         end
       end
     end