+++ db/migrate/20140704110631_add_confirmed_by_to_encoded_files.rb
@@ -1,0 +1,6 @@
+class AddConfirmedByToEncodedFiles < ActiveRecord::Migration
+  def change
+    add_reference :pageflow_video_files, :confirmed_by
+    add_reference :pageflow_audio_files, :confirmed_by
+  end
+end